Relocate gin authentication middleware
